requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
django-treebeard==4.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sqlparse==0.5.0
typing-extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
gunicorn==22.0.0
